Change overriden entities from Sylius to be traits
