--- src/Resources/config/doctrine/Adjustment.orm.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Paweł Jędrzejewski
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<!-- TODO Remove this mapping after update to Sylius 1.9. Duplicated in order to provide proper adjustments handling. -->
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping">
-
-    <mapped-superclass name="Sylius\RefundPlugin\Entity\Adjustment" table="sylius_adjustment">
-        <field name="details" type="json" />
-
-        <many-to-one field="shipment" target-entity="Sylius\RefundPlugin\Entity\ShipmentInterface" inversed-by="adjustments">
-            <join-column name="shipment_id" on-delete="CASCADE"/>
-        </many-to-one>
-    </mapped-superclass>
-
-</doctrine-mapping>
