--- src/Resources/config/doctrine/Shipment.orm.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Paweł Jędrzejewski
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<!-- TODO Remove this mapping after update to Sylius 1.9. Duplicated in order to provide proper adjustments handling. -->
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd"
->
-
-    <mapped-superclass name="Sylius\RefundPlugin\Entity\Shipment" table="sylius_shipment">
-        <field name="adjustmentsTotal" column="adjustments_total" type="integer" />
-
-        <one-to-many field="adjustments" target-entity="Sylius\RefundPlugin\Entity\AdjustmentInterface" mapped-by="shipment" orphan-removal="true">
-            <cascade>
-                <cascade-all/>
-            </cascade>
-        </one-to-many>
-    </mapped-superclass>
-
-</doctrine-mapping>
